Move supported samples detection to a separate function
